+++ test/integration/025_duplicate_model_test/models-5/subdir1/model_alt.sql
@@ -1,0 +1,1 @@
+select 100 as somevalue