+++ test/integration/025_duplicate_model_test/models-5/subdir2/model_alt.sql
@@ -1,0 +1,4 @@
+{{
+  config(enabled = False)
+}}
+select 100 as altvalue