+++ test/integration/025_duplicate_model_test/models-5/subdir3/model_alt.sql
@@ -1,0 +1,5 @@
+{{
+  config(enabled = False)
+}}
+
+select 100 as thirdvalue